requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
langchain-pinecone
pinecone-client
python-dotenv
pypdf
